requirements.txt
PyYAML~=6.0.2

# --- LangChain 核心与扩展 ---
langchain==0.3.14
langchain-core>=0.1.0
langchain-openai==0.3.0
langchain-google-vertexai
langchain-community

# --- Crawl
undetected_chromedriver

# --- Logger
loguru